+++ src/controllers/transfer/transfer.test.ts
@@ -7,7 +7,7 @@ import { DEFAULT_ACCOUNT_LABEL } from '../../consts/account'
 import { FEE_COLLECTOR } from '../../consts/addresses'
 import { networks } from '../../consts/networks'
 import { TokenResult } from '../../libs/portfolio'
-import * as ensDomainsModule from '../../services/ensDomains'
+import * as ensDomainsModule from '../../services/ensDomains/ensDomains'
 import { DomainsController } from '../domains/domains'
 import { TransferController } from './transfer'
